back-end/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
passlib[bcrypt]
python-dotenv
